api_yamdb/requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.0.12
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-filter==22.1
django-import-export==3.0.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
gunicorn==20.0.4
psycopg2-binary==2.9.6
et-xmlfile==1.1.0
flake8==5.0.4
idna==3.4
iniconfig==2.0.0
MarkupPy==1.14
mccabe==0.7.0
odfpy==1.4.1
openpyxl==3.1.1
packaging==23.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.3
tablib==3.3.0
toml==0.10.2
typing_extensions==4.4.0
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0
zipp==3.13.0
